% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyBB_CR.R
\name{AgeProRun}
\alias{AgeProRun}
\title{Iteratively apply agepro}
\usage{
AgeProRun(direct = NULL, proj.fname = NULL, FracBmsyThreshHi = NULL,
  FracBmsyThreshLo = NULL, FracFtarg = NULL, Bmsy = NULL,
  Fmsy = NULL, decimals = NULL, domsy = NULL, msy.name = NULL,
  CIwantLow = NULL, CIwantHi = NULL, fmsyold = NULL,
  Bmsyold = NULL, msyold = NULL, recrold = NULL, nyr.avg = NULL)
}
\arguments{
\item{direct}{is the directory location of proj.fname}

\item{proj.fname}{is the age pro input file without the .INP at the end}

\item{FracBmsyThreshHi}{is the fraction of Bmsy above which fishing is held constant at FracFtarg.  F changes linearly between FracBmsyThreshLo and FracBmsyThreshHi}

\item{FracBmsyThreshLo}{is the fraction of Bmsy at which no fishing occurs.}

\item{FracFtarg}{is the proportion of Fmsy that defines the maximum F desired.  Enter the fraction between 0 and 1, not the desired F rate.}

\item{Bmsy}{is Bmsy}

\item{Fmsy}{is exactly that; Fmsy.}

\item{decimals}{is the number of decimals for output values}

\item{domsy}{do longterm MSY ref point run: TRUE or FALSE}

\item{msy.name}{name of agepro run for longterm ref point calculation}

\item{fmsyold}{old fmsy value}

\item{Bmsyold}{is old Bmsy value}

\item{msyold}{is old msy value}

\item{recrold}{is old recruitment at MSY value}

\item{nyr.avg}{is the number of years from the end of the longterm projection (e.g., last 10 years) to be averaged to define ref points (e.g., Bmsy)}
}
\description{
Does iterative agepro runs
}
\examples{
AgeProRun()
}
\keyword{AgePro}
