% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_AgePro.R
\name{AnaAgePro}
\alias{AnaAgePro}
\title{analyzes age pro output}
\usage{
AnaAgePro(proj.fname.b = NULL, direct = NULL, fmsy = NULL,
  ssbmsy = NULL, decimals = NULL, fyr = NULL)
}
\arguments{
\item{proj.fname.b}{is the age pro input file without the .INP at the end}

\item{direct}{is the directory location of proj.fname.b}

\item{fmsy}{is fmsy}

\item{ssbmsy}{is ssbmsy}

\item{decimals}{is the number of decimals for output values}
}
\description{
summary stats from agepro
}
\examples{
AnaAgePro()
}
\keyword{AgePro}
