% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_AgePro.R
\name{BB_CR}
\alias{BB_CR}
\title{biomass based control rule}
\usage{
BB_CR(SSBstatus = NULL, SSBthresh = NULL, SSBthreshlo = NULL,
  FracFtarg = NULL, Fmsy = NULL, CtrlRule = NULL)
}
\arguments{
\item{FracFtarg}{is the proportion of Fmsy that defines the maximum F desired.  Enter the fraction between 0 and 1, not the desired F rate.}

\item{Fmsy}{is exactly that; Fmsy.}

\item{CtrlRule}{is a place holder for later when more control rule options are added.  Now defaults to 1, biomass based.}

\item{FracBmsyThreshLo}{is the fraction of Bmsy at which no fishing occurs.}

\item{FracBmsyThreshHi}{is the fraction of Bmsy above which fishing is held constant at FracFtarg.  F changes linearly between FracBmsyThreshLo and FracBmsyThreshHi}
}
\description{
apply a hockey stick control rule
}
\examples{
BB_CR()
}
\keyword{AgePro}
