% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_AgePro.R
\name{MSYageprofxn}
\alias{MSYageprofxn}
\title{calculates MSY ref points from longterm agepro run}
\usage{
MSYageprofxn(proj.fname.b = NULL, short.proj.name = NULL,
  direct = NULL, decimals = NULL, fmsyold = NULL, nyr.avg = NULL,
  CIwant = NULL)
}
\arguments{
\item{proj.fname.b}{is the age pro input file without the .INP at the end}

\item{short.proj.name}{is define by code}

\item{direct}{is the directory location of proj.fname.b}

\item{decimals}{is the number decimals for output values}

\item{fmsyold}{is the old value for Fmsy to be replaced by the results of this projection}

\item{nyr.avg}{is the number of years from the end of the longterm projection to be averaged to define ref points (e.g., Bmsy)}

\item{CIwant}{upper and lower quantiles to define confidence intervals}

\item{fmsy}{is fmsy}
}
\description{
MSY from agepro run
}
\examples{
OFLfxn()
}
\keyword{AgePro}
