% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_AgePro.R
\name{OFLfxn}
\alias{OFLfxn}
\title{calculates OFL based on Fmsy fishing rate}
\usage{
OFLfxn(fyr = NULL, direct = NULL, s = NULL, proj.fname = NULL,
  input = NULL, Fmsy = NULL, harvscen.num = NULL,
  harvscen.ofl = NULL)
}
\arguments{
\item{direct}{is the directory location of proj.fname.b}

\item{proj.fname.b}{is the age pro input file without the .INP at the end}

\item{fmsy}{is fmsy}
}
\description{
OFL from agepro run
}
\examples{
OFLfxn()
}
\keyword{AgePro}
